requirements.txt
bs4
colorama
dnspython
prettytable
requests
validators